+++ src/server/core/config.ts
@@ -89,6 +89,25 @@ export function readConfiguration(path: string): ConfigurationFile {
   return {};
 }
 
+function deepMerge(obj: any, value: any) {
+  Object.keys(value).forEach((key) => {
+    const oldItem = obj[key];
+    const newItem = value[key];
+
+    if (newItem === undefined) {
+      delete obj[key];
+    } else if (Array.isArray(oldItem) && Array.isArray(newItem)) {
+      obj[key] = [...oldItem, ...newItem] as any;
+    } else if (typeof oldItem === 'object') {
+      obj[key] = deepMerge({ ...oldItem }, newItem);
+    } else {
+      obj[key] = newItem;
+    }
+  });
+
+  return obj;
+}
+
 function mergeObjects<T>(
   sources: Array<Partial<KrasConfiguration>>,
   select: (config: Partial<KrasConfiguration>) => Dict<T>,
@@ -99,21 +118,7 @@ function mergeObjects<T>(
     const value = select(source);
 
     if (value && typeof value === 'object') {
-      Object.keys(value).forEach((key) => {
-        const oldItem = obj[key];
-        const newItem = value[key];
-
-        if (newItem === undefined) {
-          delete obj[key];
-        } else if (typeof oldItem === 'object') {
-          obj[key] = {
-            ...oldItem,
-            ...newItem,
-          };
-        } else {
-          obj[key] = newItem;
-        }
-      });
+      deepMerge(obj, value);
     }
   }